Move properties into their own API
