+++ betty/property/__init__.py
@@ -5,7 +5,6 @@ Object properties.
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from collections.abc import Iterable, Mapping, MutableMapping, MutableSequence
 from typing import TYPE_CHECKING, Any, Self, cast, final, overload, override
 
 from betty.data import OptionalDefinition
@@ -17,13 +16,7 @@ from betty.typing import Void
 if TYPE_CHECKING:
     from collections.abc import Callable
 
-    from ty_extensions import Intersection
-
-    from betty.collection.keyed import MutableKeyedCollection
     from betty.data import Data, DataDefinition
-    from betty.data.aggregate.collection.keyed import KeyedCollectionDefinition
-    from betty.data.aggregate.collection.mapping import MappingDefinition
-    from betty.data.aggregate.collection.sequence import SequenceDefinition
     from betty.locale.localizable import ResolvableLocalizable
 
 
@@ -129,7 +122,7 @@ class Property[ValueGetT, ValueSetT](_Property[ValueGetT, ValueSetT]):
 @final
 class PropertyNotInitialized(ValueError):
     """
-    Raised when a class failed to initialize a value for a :py:class:`betty.data.aggregate.record.object.property.Property`.
+    Raised when a class failed to initialize a value for a :py:class:`betty.property.Property`.
     """
 
 
@@ -183,144 +176,3 @@ class Optional[ValueGetT, ValueSetT](_Property[ValueGetT | None, ValueSetT | Non
         Delete the value from the instance.
         """
         self.set(instance, None)
-
-
-class MappingProperty[MutableMappingT: MutableMapping[Any, Any], ValueSetT](
-    Property[MutableMappingT, ValueSetT]
-):
-    """
-    A property that contains a :py:class:`collections.abc.MutableMapping`.
-    """
-
-    _data: MappingDefinition[MutableMappingT]
-
-    def __init__(
-        self,
-        data: Intersection[DataDefinition[MutableMappingT], MappingDefinition]
-        | Data[Intersection[DataDefinition[MutableMappingT], MappingDefinition]],
-        *,
-        label: ResolvableLocalizable | None = None,
-        description: ResolvableLocalizable | None = None,
-        omit_load: bool | None = None,
-        omit_dump: Callable[[MutableMappingT], bool] | None = None,
-        default: Callable[[], Mapping] = dict,
-    ):
-        super().__init__(
-            data,
-            label=label,
-            description=description,
-            omit_load=omit_load,
-            omit_dump=omit_dump,
-            default=self._new_default,
-        )
-        self._default_values = default
-
-    def _new_default(self) -> MutableMappingT:
-        new = self._data.new()
-        new.update(self._default_values())
-        return new
-
-    @override
-    def set(self, instance: Any, value: ValueSetT | MutableMappingT) -> MutableMappingT:
-        data = self.get(instance)
-        data.clear()
-        data.update(self._resolver(value))
-        return data
-
-
-class SequenceProperty[MutableSequenceT: MutableSequence[Any], ValueSetT](
-    Property[MutableSequenceT, ValueSetT]
-):
-    """
-    A property that contains a :py:class:`collections.abc.MutableSequence`.
-    """
-
-    _data: SequenceDefinition[MutableSequenceT]
-
-    def __init__(
-        self,
-        data: Intersection[DataDefinition[MutableSequenceT], SequenceDefinition]
-        | Data[Intersection[DataDefinition[MutableSequenceT], SequenceDefinition]],
-        *,
-        label: ResolvableLocalizable | None = None,
-        description: ResolvableLocalizable | None = None,
-        omit_load: bool | None = None,
-        omit_dump: Callable[[MutableSequenceT], bool] | None = None,
-        resolver: Callable[
-            [ValueSetT | Iterable[ValueSetT]], Iterable[ValueSetT]
-        ] = passthrough,
-        default: Callable[[], ValueSetT | Iterable[ValueSetT]] = list,
-    ):
-        super().__init__(
-            data,
-            label=label,
-            description=description,
-            omit_load=omit_load,
-            omit_dump=omit_dump,
-            default=self._new_default,
-        )
-        self._values_resolver = resolver
-        self._default_values = default
-
-    def _new_default(self) -> MutableSequenceT:
-        new = self._data.new()
-        new.extend(self._values_resolver(self._default_values()))
-        return new
-
-    @override
-    def set(
-        self, instance: Any, value: ValueSetT | MutableSequenceT
-    ) -> MutableSequenceT:
-        data = self.get(instance)
-        data.clear()
-        data.extend(self._values_resolver(value))
-        return data
-
-
-class KeyedCollectionProperty[
-    MutableKeyedCollectionT: MutableKeyedCollection,
-    ValueSetT,
-](Property[MutableKeyedCollectionT, Iterable[ValueSetT]]):
-    """
-    A property that contains an :py:class:`betty.collection.keyed.KeyedCollection`.
-    """
-
-    _data: KeyedCollectionDefinition[MutableKeyedCollectionT]
-
-    def __init__(
-        self,
-        data: KeyedCollectionDefinition[MutableKeyedCollectionT],
-        *,
-        label: ResolvableLocalizable | None = None,
-        description: ResolvableLocalizable | None = None,
-        omit_load: bool | None = None,
-        omit_dump: Callable[[MutableKeyedCollectionT], bool] | None = None,
-        resolver: Callable[
-            [ValueSetT | Iterable[ValueSetT]], Iterable[ValueSetT]
-        ] = passthrough,
-        default: Callable[[], ValueSetT | Iterable[ValueSetT]] = list,
-    ):
-        super().__init__(
-            data,
-            label=label,
-            description=description,
-            omit_load=omit_load,
-            omit_dump=omit_dump,
-            default=self._new_default,
-        )
-        self._values_resolver = resolver
-        self._default_values = default
-
-    def _new_default(self) -> MutableKeyedCollectionT:
-        new = self._data.new()
-        new.add(*self._values_resolver(self._default_values()))
-        return new
-
-    @override
-    def set(
-        self, instance: Any, value: Iterable[ValueSetT] | MutableKeyedCollectionT
-    ) -> MutableKeyedCollectionT:
-        data = self.get(instance)
-        data.clear()
-        data.add(*self._values_resolver(value))
-        return data
