+++ betty/property/collection/__init__.py
@@ -1,0 +1,3 @@
+"""
+Collection properties.
+"""
